Generalize implementation
