+++ package.json
@@ -9,6 +9,7 @@
   "scripts": {
     "prepublish": "npm run build",
     "build": "node scripts/build.js && cp derivable.d.ts derivable.js.flow dist/",
+    "pretest": "npm run build",
     "test": "jest",
     "ci:test": "jest && (cd test_flow && npm install && npm test)",
     "bench": "node scripts/bench.js",
